+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unrelased]
 
+## [0.5.2d] - 09/06/2020
+### Changed
+- `load_env()` function in `mains` module. Added kwargs forwarding. If dotenv_path or stream is present it will be used.
+Otherwise, dotenv_path will be constructed from ENV_PCK and ENV_NAME.  
+
+
 ## [0.5.2c] - 19/05/2020
 ### Added 
 - `path()` function in `mains` module. For older Python version uses